Web: PortalSettings: Restore: Adding a new component.
